upd: improve User repository and class
