+++ src/ImpoBooks/ImpoBooks.DataAccess/Interfaces/IUsersRepository.cs
@@ -1,0 +1,8 @@
+using ImpoBooks.DataAccess.Entities;
+
+namespace ImpoBooks.DataAccess.Interfaces;
+
+public interface IUsersRepository : IRepository<User>
+{
+    Task<User> GetByEmailAsync(string email);
+}
